# event-api/docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  models.RoleRequestApprovalRequest:
    properties:
      notes:
        maxLength: 500
        type: string
      request_id:
        type: string
    required:
    - request_id
    type: object
  models.RoleRequestRejectionRequest:
    properties:
      reason:
        maxLength: 500
        minLength: 10
        type: string
      request_id:
        type: string
    required:
    - reason
    - request_id
    type: object
  models.RoleRequestResponse:
    properties:
      message:
        type: string
      status:
        description: '"pending", "approved", "rejected"'
        type: string
    type: object
...
  models.RoleRequestsListResponse:
    properties:
      requests:
        items:
          $ref: '#/definitions/models.RoleRequestStatus'
        type: array
      total:
        type: integer
    type: object
  models.SubscribeRequest:
    properties:
      metadata:
        additionalProperties:
          type: string
        example:
          dietary_preferences: vegan
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления событиями и аутентификации пользователей
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Event API
  version: "1.0"
paths:
  /api/admin/role-requests/{requestId}/approve:
    post:
      consumes:
      - application/json
      description: Approve a pending role request (admin only).
      parameters:
      - description: Approval request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RoleRequestApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message and status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve role request
      tags:
      - admin
  /api/admin/role-requests/{requestId}/reject:
    post:
      consumes:
      - application/json
      description: Reject a pending role request (admin only).
      parameters:
      - description: Rejection request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RoleRequestRejectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message and status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject role request
      tags:
      - admin
  /api/admin/role-requests/pending:
    get:
      consumes:
      - application/json
      description: Get all pending role requests (admin only).
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoleRequestsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pending role requests
      tags:
      - admin
  /api/events/{event_id}/participants:
    get:
      consumes:
      - application/json
      description: Returns a list of confirmed participants for a specific event.
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Participant'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event participants
      tags:
      - events
  /api/users/me:
    get:
      consumes:
      - application/json
      description: Returns full user profile and status flags.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /api/users/me/events/{event_id}/subscribe:
    post:
      consumes:
      - application/json
      description: Idempotent subscription endpoint. Checks capacity and handles waitlisting.
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Subscription metadata
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.SubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Already subscribed
          schema:
            $ref: '#/definitions/models.EventSubscription'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EventSubscription'
        "202":
          description: Waitlisted
          schema:
            $ref: '#/definitions/models.EventSubscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Event full
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Subscribe to an event
      tags:
      - users
  /api/users/me/events/history:
    get:
      consumes:
      - application/json
      description: Returns attended or missed past events.
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventWithSubscription'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event history
      tags:
      - users
  /api/users/me/events/upcoming:
    get:
      consumes:
      - application/json
      description: Returns events the user is currently subscribed to (confirmed participation).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventWithSubscription'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get upcoming events
      tags:
      - users
  /api/users/public/{userId}:
    get:
      consumes:
      - application/json
      description: Returns public information about a user by their UUID. Does not
        require authentication.
      parameters:
      - description: User ID (UUID format)
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public user profile
          schema:
            $ref: '#/definitions/models.PublicUserProfile'
        "304":
          description: Not Modified - resource unchanged since last request
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found or profile is private
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get public user profile
      tags:
      - users
  /api/users/request-role:
    post:
      consumes:
      - application/json
      description: User requests a role upgrade (creator, support, etc.) with a reason.
      parameters:
      - description: Role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoleRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request role upgrade
      tags:
      - users
  /api/users/request-role/all:
    get:
      consumes:
      - application/json
      description: Get all role requests for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RoleRequestStatus'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all role requests
      tags:
      - users
  /api/users/request-role/status:
    get:
      consumes:
      - application/json
      description: Get the status of a specific role request.
      parameters:
      - description: Role
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoleRequestStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role request status
      tags:
      - users
...
  /v1/api/admin/events/{eventId}/block:
    post:
      consumes:
      - application/json
      description: Переносит событие в заблокированные с указанием причины
      parameters:
      - description: ID события
        in: path
        name: eventId
        required: true
        type: string
      - description: Причина блокировки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BlockEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Нет авторизации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Заблокировать событие
      tags:
      - admin
  /v1/api/admin/events/{eventId}/comments:
    get:
      description: Возвращает историю комментариев модерации для указанного события
        без проверки владельца
      parameters:
      - description: ID события
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReviewComment'
            type: array
        "401":
          description: Нет авторизации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить комментарии к событию (админ)
      tags:
      - admin
  /v1/api/admin/events/{eventId}/request-revision:
    post:
      consumes:
      - application/json
      description: Переводит событие в статус needs_revision с комментарием
      parameters:
      - description: ID события
        in: path
        name: eventId
        required: true
        type: string
      - description: Комментарий
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddReviewCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Нет авторизации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Запросить доработку события
      tags:
      - admin
  /v1/api/admin/users:
    get:
      description: Возвращает список всех пользователей системы
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список всех пользователей
      tags:
      - admin
  /v1/api/admin/users/role:
    put:
      consumes:
      - application/json
      description: 'Назначает пользователю одну из ролей: `user`, `creator`, `support`,
        `admin`. Только для администраторов.'
      parameters:
      - description: 'Данные для обновления роли (role: user|creator|support|admin)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный пользователь
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверный формат запроса или недопустимая роль
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить роль пользователя
      tags:
      - admin
  /v1/api/auth/check-user:
    post:
      consumes:
      - application/json
      description: Проверяет, существует ли пользователь с указанным email и/или телефоном.
        Используется на этапе регистрации для предотвращения дублирования.
      parameters:
      - description: Email и/или телефон для проверки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CheckUserExistsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Результат проверки
          schema:
            $ref: '#/definitions/models.CheckUserExistsResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Проверка существования пользователя
      tags:
      - auth
  /v1/api/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает JWT токены
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /v1/api/auth/logout:
    post:
      consumes:
      - application/json
      description: Инвалидирует refresh токен пользователя
      parameters:
      - description: Refresh токен для инвалидации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - auth
  /v1/api/auth/me:
    get:
      description: 'Возвращает данные текущего авторизованного пользователя с информацией
        о статусе регистрации в Telegram и ролью пользователя (role: user|creator|support|admin)'
      produces:
      - application/json
      responses:
        "200":
          description: 'Информация о пользователе с полями: user (models.User), role
            (string), telegram_registered (bool), telegram_info (object, опционально)'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить информацию о текущем пользователе
      tags:
      - auth
  /v1/api/auth/register:
    post:
      consumes:
      - application/json
      description: 'Создает нового пользователя и отправляет код верификации через
        email и/или SMS (в зависимости от verification_type: "email", "sms", "both")'
      parameters:
      - description: 'Данные для регистрации (verification_type опционален: email/sms/both,
          по умолчанию both)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user:models.User, verify_code:string
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
  /v1/api/auth/resend-code:
    post:
      consumes:
      - application/json
      description: Генерирует и отправляет новый код верификации через выбранный канал
        (email/telegram/sms)
      parameters:
      - description: Email и тип верификации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ResendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Код успешно отправлен
          schema:
            $ref: '#/definitions/models.ResendCodeResponse'
        "400":
          description: Неверный формат запроса или невалидные данные
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден или уже верифицирован
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Превышен лимит запросов
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при отправке кода
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Повторная отправка кода верификации
      tags:
      - auth
  /v1/api/auth/verify:
    post:
      consumes:
      - application/json
      description: Проверяет код верификации для подтверждения email и возвращает
        JWT токен
      parameters:
      - description: Код верификации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email верифицирован, токен выдан
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверный код или формат запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Верификация email
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"